requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
jinja2
python-jose[cryptography]
pydantic-settings
passlib[bcrypt]
